Break out bad ideas to individual cogs
