+++ big_name/__init__.py
@@ -1,0 +1,4 @@
+from .big_name import BigName
+
+def setup(bot):
+    bot.add_cog(BigName(bot))
