+++ clone/__init__.py
@@ -1,0 +1,4 @@
+from .clone import Clone
+
+def setup(bot):
+    bot.add_cog(Clone(bot))
