+++ clone/info.json
@@ -1,0 +1,4 @@
+{
+  "author" : ["Eris", "Maddie"],
+  "description" : "Usage: [p]clone <User>",
+  "name" : "The bot takes on a specified user's name and avatar"}
