+++ weave/info.json
@@ -1,0 +1,4 @@
+{
+  "author" : ["Eris", "Maddie"],
+  "description" : "Usage: [p]weave <Emoji> <Emoji> <width> <height>",
+  "name" : "Create a mosaic of alternating specified emoji. Emojis may be animated or still"}
